refactored new function into tools.c file
