+++ MANIFEST.in
@@ -1,9 +1,11 @@
 include src/assist.c
 include src/forces.c
+include src/tools.c
 include src/planets.c
 include src/spk.c
 include src/assist.h
 include src/forces.h
+include src/tools.h
 include src/planets.h
 include src/spk.h
 include README.md
