+++ src/tools.h
@@ -1,0 +1,7 @@
+#ifndef _TOOLS_H
+#define _TOOLS_H
+#include "assist.h"
+
+
+#endif // _TOOLS_H
+
